ref(router): Inline the legacy stories redirect
