+++ static/app/router/routes.spec.tsx
@@ -1,7 +1,9 @@
 import {matchRoutes, type RouteObject} from 'react-router-dom';
 
+import {render, waitFor} from 'sentry-test/reactTestingLibrary';
+
 import * as constants from 'sentry/constants';
-import {buildRoutes} from 'sentry/router/routes';
+import {buildRoutes, LegacyStoriesRedirect} from 'sentry/router/routes';
 import {replaceRouterParams} from 'sentry/utils/replaceRouterParams';
 import {normalizeUrl} from 'sentry/utils/url/normalizeUrl';
 
@@ -183,5 +185,25 @@ describe('buildRoutes()', () => {
       '/:orgId/:projectId/',
       'events/:eventId/',
     ]);
+  });
+});
+
+describe('LegacyStoriesRedirect', () => {
+  it('redirects stories deep links to scraps and preserves the query and hash', async () => {
+    const {router} = render(<LegacyStoriesRedirect />, {
+      initialRouterConfig: {
+        location: {
+          pathname: '/organizations/test-org/stories/core/button/?theme=dark#examples',
+        },
+      },
+    });
+
+    await waitFor(() => {
+      expect(router.location.pathname).toBe(
+        '/organizations/test-org/scraps/core/button/'
+      );
+    });
+    expect(router.location.query).toEqual({theme: 'dark'});
+    expect(router.location.hash).toBe('#examples');
   });
 });
