--- static/app/stories/view/legacyStoriesRedirect.spec.tsx
@@ -1,23 +1,0 @@
-import {render, waitFor} from 'sentry-test/reactTestingLibrary';
-
-import LegacyStoriesRedirect from 'sentry/stories/view/legacyStoriesRedirect';
-
-describe('LegacyStoriesRedirect', () => {
-  it('redirects stories deep links to scraps and preserves the query and hash', async () => {
-    const {router} = render(<LegacyStoriesRedirect />, {
-      initialRouterConfig: {
-        location: {
-          pathname: '/organizations/test-org/stories/core/button/?theme=dark#examples',
-        },
-      },
-    });
-
-    await waitFor(() => {
-      expect(router.location.pathname).toBe(
-        '/organizations/test-org/scraps/core/button/'
-      );
-    });
-    expect(router.location.query).toEqual({theme: 'dark'});
-    expect(router.location.hash).toBe('#examples');
-  });
-});
