--- static/app/stories/view/legacyStoriesRedirect.tsx
@@ -1,18 +1,0 @@
-import {Navigate} from 'react-router-dom';
-
-import {useLocation} from 'sentry/utils/useLocation';
-
-export default function LegacyStoriesRedirect() {
-  const location = useLocation();
-
-  return (
-    <Navigate
-      replace
-      to={{
-        pathname: location.pathname.replace(/\/stories(?=\/|$)/, '/scraps'),
-        search: location.search,
-        hash: location.hash,
-      }}
-    />
-  );
-}
